Add logger and format code with yapf

+++ src/storage.py
@@ -1,11 +1,15 @@
 import sqlite3
-import time
 from sqlite3 import Error
+
+from bot_logger import get_logger
+
+_logger = get_logger(__name__)
 
 
 # ------------------------------------------------------------
 class ParseError(Exception):
     pass
+
 
 # Пример кортежа идущего в БД на вставку
 # (123456, "test.dev.domain.ru", "Тестовый сертификат", '2023-06-10')
@@ -34,8 +38,7 @@ class CertModel:
         if len(parts) != 3:
             raise ParseError("Не удалось разобрать данные: " + cert_str)
 
-        return CertModel(
-            (None, None, parts[1].strip(), parts[2].strip(), parts[0].strip()))
+        return CertModel((None, None, parts[1].strip(), parts[2].strip(), parts[0].strip()))
 
 
 # ------------------------------------------------------------
@@ -67,47 +70,43 @@ class CertStore:
     def add_cert(self, cert):
         try:
             cur = self.conn.cursor()
-            cur.execute(
-                f"insert into cert( {CertStore._INSERT_DB_FIELDS_STR} ) values(?,?,?,?);",
-                (cert.user_id, cert.cn, cert.description, cert.exp_date))
+            cur.execute(f"insert into cert( {CertStore._INSERT_DB_FIELDS_STR} ) values(?,?,?,?);",
+                        (cert.user_id, cert.cn, cert.description, cert.exp_date))
             self.conn.commit()
             cert.id = cur.lastrowid
         except Error as e:
-            print(e)
-            raise e
+            _logger.exception("Ошибка записи в БД: " + str(e))
+            raise
         return cert
 
     def delete_cert(self, user_id, cert_id):
         try:
             cur = self.conn.cursor()
-            cur.execute("delete from cert where id = ? and user_id = ?;",
-                        (cert_id, user_id))
+            cur.execute("delete from cert where id = ? and user_id = ?;", (cert_id, user_id))
             self.conn.commit()
         except Error as e:
-            print(e)
-            raise e
+            _logger.exception("Ошибка удаления записи из БД: " + str(e))
+            raise
 
     def get_cert(self, user_id, cert_id):
         try:
             cur = self.conn.cursor()
-            cur.execute(
-                "select id, user_id, cn, description, exp_date from cert where id = ? and user_id = ?;",
-                (cert_id, user_id))
+            cur.execute("select id, user_id, cn, description, exp_date from cert where id = ? and user_id = ?;",
+                        (cert_id, user_id))
             rows = cur.fetchall()
         except Error as e:
-            print(e)
-            raise e
+            _logger.exception("Ошибка получения записи из БД: " + str(e))
+            raise
         return CertStore.rows_to_cert_model(rows)
 
     def find_by_cn(self, user_id, common_name):
         try:
             cur = self.conn.cursor()
-            cur.execute(
-                "select id, user_id, cn, description, exp_date from cert where user_id = ? and cn = ?;",
-                (user_id, common_name))
+            cur.execute("select id, user_id, cn, description, exp_date from cert where user_id = ? and cn = ?;",
+                        (user_id, common_name))
             rows = cur.fetchall()
         except Error as e:
-            print(e)
+            _logger.exception("Ошибка поиска записи в БД: " + str(e))
             raise e
         return CertStore.rows_to_cert_model(rows)
